Directory Management
This introduces a directory abstraction and creates a local
implementation of this abstraction. I considered creating a separate
gem for the local implementation, but concluded that it would be
easier for people to understand the core library if there was a
reference implementation that was easy to find. This isn't a big
problem, either, as it doesn't require any additional gems to interact
with local filesystems.