Decompose Write Abstraction and Support Subdirectory Contents
I discovered in working on one of the other libraries that I neglected
to add support for copying the actual files in the directory being
written. This forced the introduction of a general `files` instance
variable, along with an accessor that implementations can override.

Much like `find`, `write` had some general structure that we want to
use to limit what needs to be implemented to support writing.