Common Path and Equality Management
The `path` value has shown up in every single implementation of the
abstraction thus far. While it might not always be present, it's
common enough to pull into the parent implementation as an alternative
to `name`. This also allows me to pull up the `path_of`
implementation.

With much of what's compared remaining the same between
implementations, it made sense to pull equality up to the parent
class. When I did so, however, I realized that `==` made more sense as
a comparison between files of the directory and `eql?` more sense when
comparing the identity.