Move file writing to File class
I realized that it made more sense for the responsibility for writing
a file to live in the File class instead of Directory. This may seem
like a small thing, but it opens the potential of people using the
File class without needing to engage in the Directory class if they
don't need that code.